time of the day added to elapsed time in ts
training_ts had timestamps starting from 0 in hh:mm:ss format. This is now converted into actual time of the day.

When merging the gps wiht the ts data a lot of the values did not match. There seems to be an error with the time measurement (used as keys) between the two